v0.0.1 completed:start dual linux kernels. (#28)
* start 2 linux successfully, close #24

+++ src/device/pci.rs
@@ -3,7 +3,7 @@ use crate::{
     memory::{mmio_perform_access, MMIOAccess},
 };
 
-pub fn mmio_pci_handler(mmio: &mut MMIOAccess, base: u64) -> HvResult {
+pub fn mmio_pci_handler(mmio: &mut MMIOAccess, base: usize) -> HvResult {
     mmio_perform_access(base, mmio);
     Ok(())
 }
